backend/cloudfunctions/requirements.txt
google-cloud-firestore==2.0.2